Create separate modules for AnswerService, QuestionGraphService, and SessionService
