+++ api/app/signals/apps/questionnaires/services/__init__.py
@@ -1,12 +1,18 @@
 # SPDX-License-Identifier: MPL-2.0
 # Copyright (C) 2019 - 2021 Gemeente Amsterdam
 
+from signals.apps.questionnaires.services.answer import AnswerService
 from signals.apps.questionnaires.services.feedback_request import FeedbackRequestService
+from signals.apps.questionnaires.services.question_graph import QuestionGraphService
 from signals.apps.questionnaires.services.questionnaires import QuestionnairesService
 from signals.apps.questionnaires.services.reaction_request import ReactionRequestService
+from signals.apps.questionnaires.services.session import SessionService
 
 __all__ = [
     'FeedbackRequestService',
     'QuestionnairesService',
     'ReactionRequestService',
+    'AnswerService',
+    'QuestionGraphService',
+    'SessionService',
 ]
